Move ReviewingSettingsFragment to a new file
